Add files via upload
Added functionality to set file with list of putative paralogs to remove for all samples in step dataset optimization 2a. 
This is useful if only a subset taxa is used and the paralog detection from another analysis should be applied. 
E.g. this can be used for phased accessions to remove the same paralogs as detected in the non phased accessions.